+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 Bug fixes and improvements (backwards-compatible):
 * [OLMIS-2871](https://openlmis.atlassian.net/browse/OLMIS-2871): The service now uses an Authorization header instead of an access_token request parameter when communicating with other services.
+* [OLMIS-3059](https://openlmis.atlassian.net/browse/OLMIS-3059): The search orders endpoint now sorts the orders by created date property (most recent first).
 
 5.0.0 / 2017-07-20
 ==================